requirements.txt
torch>=2.4.1
transformers>=4.45.1
pillow>=10.4.0
